refactor: split homepage controller into specialized domain controllers
- Extract distribution logic into new distribution_controller.js
- Extract mining metrics into new mining_controller.js
- Extract voting logic into new voting_controller.js
- Create ska_helper.js for SKA reward formatting utilities
- Simplify homepage_controller.js to handle only core homepage concerns
- Update home.tmpl and home_voting.tmpl to use new controllers via data-controller attributes
- Reduce homepage_controller targets from 40+ to 6 core targets
- Improve code organization and maintainability by separating concerns
