+++ cmd/dcrdata/public/js/controllers/homepage_controller.js
@@ -36,44 +36,7 @@ function mempoolTableRow(tx) {
 
 export default class extends Controller {
   static get targets() {
-    return [
-      'transactions',
-      'difficulty',
-      'bsubsidyTotal',
-      'bsubsidyPow',
-      'bsubsidyPos',
-      'bsubsidyDev',
-      'coinSupply',
-      'blocksdiff',
-      'devFund',
-      'windowIndex',
-      'posBar',
-      'rewardIdx',
-      'powBar',
-      'poolSize',
-      'poolValue',
-      'ticketReward',
-      'targetPct',
-      'poolSizePct',
-      'hashrate',
-      'hashrateDelta',
-      'nextExpectedSdiff',
-      'nextExpectedMin',
-      'nextExpectedMax',
-      'mempool',
-      'voteTally',
-      'powConverted',
-      'convertedDev',
-      'convertedSupply',
-      'convertedDevSub',
-      'exchangeRate',
-      'convertedStake',
-      'mixedPct',
-      'indicatorList',
-      'totalBar',
-      'skaVoteRewards',
-      'coinFillBars'
-    ]
+    return ['transactions', 'mempool', 'voteTally', 'indicatorList', 'totalBar', 'coinFillBars']
   }
 
   connect() {
@@ -203,135 +166,7 @@ export default class extends Controller {
   }
 
   _processBlock(blockData) {
-    const ex = blockData.extra
-    this.difficultyTarget.innerHTML = humanize.threeSigFigs(ex.difficulty)
-    this.bsubsidyPowTarget.innerHTML = humanize.decimalParts(
-      ex.subsidy.pow / 100000000,
-      false,
-      8,
-      2
-    )
-    this.bsubsidyPosTarget.innerHTML = humanize.decimalParts(
-      ex.subsidy.pos / 500000000,
-      false,
-      8,
-      2
-    ) // 5 votes per block (usually)
-    this.bsubsidyDevTarget.innerHTML = humanize.decimalParts(
-      ex.subsidy.dev / 100000000,
-      false,
-      8,
-      2
-    )
-    this.coinSupplyTarget.innerHTML = humanize.decimalParts(ex.coin_supply / 100000000, true, 0)
-    this.mixedPctTarget.innerHTML = ex.mixed_percent.toFixed(0)
-    this.blocksdiffTarget.innerHTML = humanize.decimalParts(ex.sdiff, false, 8, 2)
-    this.nextExpectedSdiffTarget.innerHTML = humanize.decimalParts(
-      ex.next_expected_sdiff,
-      false,
-      2,
-      2
-    )
-    this.nextExpectedMinTarget.innerHTML = humanize.decimalParts(ex.next_expected_min, false, 2, 2)
-    this.nextExpectedMaxTarget.innerHTML = humanize.decimalParts(ex.next_expected_max, false, 2, 2)
-    this.windowIndexTarget.textContent = ex.window_idx
-    this.posBarTarget.style.width = `${(ex.window_idx / ex.params.window_size) * 100}%`
-    this.poolSizeTarget.innerHTML = humanize.decimalParts(ex.pool_info.size, true, 0)
-    this.targetPctTarget.textContent = parseFloat(ex.pool_info.percent_target - 100).toFixed(2)
-    this.rewardIdxTarget.textContent = ex.reward_idx
-    this.powBarTarget.style.width = `${(ex.reward_idx / ex.params.reward_window_size) * 100}%`
-    this.poolValueTarget.innerHTML = humanize.decimalParts(ex.pool_info.value, true, 0)
-    this.ticketRewardTarget.innerHTML = `${ex.reward.toFixed(2)}%`
-    this.poolSizePctTarget.textContent = parseFloat(ex.pool_info.percent).toFixed(2)
-    const treasuryTotal = ex.dev_fund + ex.treasury_bal.balance
-    this.devFundTarget.innerHTML = humanize.decimalParts(treasuryTotal / 100000000, true, 0)
-    this.hashrateTarget.innerHTML = humanize.decimalParts(ex.hash_rate, false, 8, 2)
-    this.hashrateDeltaTarget.innerHTML = humanize.fmtPercentage(ex.hash_rate_change_month)
-    this._renderSkaRewards(ex.ska_vote_rewards)
-
-    if (ex.exchange_rate) {
-      const xcRate = ex.exchange_rate.value
-      const index = ex.exchange_rate.index
-      if (this.hasPowConvertedTarget) {
-        this.powConvertedTarget.textContent = `${humanize.twoDecimals((ex.subsidy.pow / 1e8) * xcRate)} ${index}`
-      }
-      if (this.hasConvertedDevTarget) {
-        this.convertedDevTarget.textContent = `${humanize.threeSigFigs((treasuryTotal / 1e8) * xcRate)} ${index}`
-      }
-      if (this.hasConvertedSupplyTarget) {
-        this.convertedSupplyTarget.textContent = `${humanize.threeSigFigs((ex.coin_supply / 1e8) * xcRate)} ${index}`
-      }
-      if (this.hasConvertedDevSubTarget) {
-        this.convertedDevSubTarget.textContent = `${humanize.twoDecimals((ex.subsidy.dev / 1e8) * xcRate)} ${index}`
-      }
-      if (this.hasExchangeRateTarget) {
-        this.exchangeRateTarget.textContent = humanize.twoDecimals(xcRate)
-      }
-      if (this.hasConvertedStakeTarget) {
-        this.convertedStakeTarget.textContent = `${humanize.twoDecimals(ex.sdiff * xcRate)} ${index}`
-      }
-    }
-  }
-
-  // _renderSkaRewards clones the ska-reward-block-template for each entry and
-  // fills the decimal-parts spans to match the SSR decimalParts 4-element structure:
-  // int="0.16", decimal="158359431255151", trailing-zeroes="0"
-  _renderSkaRewards(rewards) {
-    if (!this.hasSkaVoteRewardsTarget) return
-    const tmpl = document.getElementById('ska-reward-block-template')
-    if (!tmpl) return
-
-    const container = this.skaVoteRewardsTarget
-    container.innerHTML = ''
-
-    rewards.forEach((r) => {
-      const clone = document.importNode(tmpl.content, true)
-
-      // Replicate skaSplitParts(s, 2): split at dot, strip trailing zeros,
-      // take first 2 significant decimals as "bold", rest as "rest".
-      const s = r.per_block || ''
-      const isDash = !s
-
-      const decimalPartsEl = clone.querySelector('.decimal-parts')
-      if (decimalPartsEl && isDash) {
-        decimalPartsEl.style.display = 'none'
-      }
-      const dot = s.indexOf('.')
-      const intPart = dot >= 0 ? s.slice(0, dot) : s
-      const frac = dot >= 0 ? s.slice(dot + 1) : ''
-      const trimmed = frac.replace(/0+$/, '')
-      const trailingZeros = frac.slice(trimmed.length)
-      const bold = trimmed.slice(0, 2)
-      const rest = trimmed.slice(2)
-
-      // Match decimalParts 4-element branch exactly:
-      // <span class="int">intPart.bold</span>
-      // <span class="decimal">rest</span>
-      // <span class="decimal trailing-zeroes">trailingZeros</span>
-      const intEl = clone.querySelector('.int')
-      const decEl = clone.querySelector('.decimal:not(.trailing-zeroes)')
-      const trailEl = clone.querySelector('.trailing-zeroes')
-
-      if (intEl) intEl.textContent = bold ? `${intPart}.${bold}` : intPart
-      if (decEl) decEl.textContent = rest
-      if (trailEl) trailEl.textContent = trailingZeros
-
-      // Fill by data-field, then strip the attributes so the rendered HTML
-      // matches the SSR output exactly.
-      clone.querySelectorAll('[data-field]').forEach((el) => {
-        const field = el.dataset.field
-        if (field === 'unit') {
-          el.textContent = isDash
-            ? `— ${r.symbol}/VAR per last block`
-            : `${r.symbol}/VAR per last block`
-        } else if (field === 'per30d') {
-          el.textContent = `${r.per_30_days} ${r.symbol}/VAR per 30 days`
-        } else if (field === 'peryear') el.textContent = `${r.per_year} ${r.symbol}/VAR per year`
-        el.removeAttribute('data-field')
-      })
-
-      container.appendChild(clone)
-    })
+    this.dispatch('block', { detail: blockData })
   }
 
   // ─── Indicator update methods ───────────────────────────────────────────────
